Fix demo

+++ python/demo/demo_axis/demo_axis.py
@@ -644,7 +644,7 @@ if has_vtx:
     v_dg_el = element("DG", msh.basix_cell(), degree, shape=(3, ))
     W = fem.functionspace(msh, v_dg_el)
     Es_dg = fem.Function(W, dtype=complex_dtype)
-    Es_expr = fem.Expression(Esh, W.element.interpolation_points())
+    Es_expr = fem.Expression(Esh, W.element.interpolation_points(), dtype=complex_dtype)
     Es_dg.interpolate(Es_expr)
     with VTXWriter(msh.comm, "sols/Es.bp", Es_dg) as f:
         f.write(0.0)
